requirements.txt
glfw
numpy
Pillow
PyOpenGL
PyOpenGL-accelerate
